myshop/requirements.txt
Django==4.1.4
djangorestframework
djangorestframework-xml
celery 
Pillow
